fix: long으로 수정

See merge request s11-webmobile1-sub2/S11P12A707!142

+++ Backend/top/src/main/java/com/ssafy/top/appfocustimes/dto/response/FocusTimeListDayResponse.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 public class FocusTimeListDayResponse {
 
     @Schema(description = "단위 시간")
-    private int startTime;
+    private Long startTime;
 
     @Schema(description = "집중 시간")
     private Long focusTime;
